Network Module: Move FeedImageDataMapper to its own file
