+++ EssentialFeed/EssentialFeed/Feed API/RemoteFeedImageDataLoader.swift
@@ -11,18 +11,3 @@ public extension RemoteFeedImageDataLoader {
 		self.init(client: client, mapper: FeedImageDataMapper.map)
 	}
 }
-
-public final class FeedImageDataMapper {
-
-	public enum Error: Swift.Error {
-		case invalidData
-	}
-
-	public static func map(_ data: Data, from response: HTTPURLResponse) throws -> Data {
-		guard response.isOK, !data.isEmpty else {
-			throw Error.invalidData
-		}
-
-		return data
-	}
-}
